+++ src/winusb/messenger-winusb.h
@@ -26,7 +26,7 @@ namespace librealsense
             usb_messenger_winusb(const std::shared_ptr<usb_device_winusb> device, std::shared_ptr<handle_winusb> handle);
             virtual ~usb_messenger_winusb();
 
-            virtual usb_status control_transfer(int request_type, int request, int value, int index, uint8_t* buffer, uint32_t length, uint32_t& transferred, uint32_t timeout_ms) override;
+            virtual usb_status control_transfer(uint8_t request_type, uint8_t request, uint16_t value, uint16_t index, uint8_t* buffer, uint16_t length, uint32_t& transferred, uint32_t timeout_ms) override;
             virtual usb_status bulk_transfer(const rs_usb_endpoint& endpoint, uint8_t* buffer, uint32_t length, uint32_t& transferred, uint32_t timeout_ms) override;
             virtual usb_status reset_endpoint(const rs_usb_endpoint& endpoint, uint32_t timeout_ms) override;
             virtual usb_status submit_request(const rs_usb_request& request) override;
